project/backend/requirements.txt
celery==5.3.0
channels==4.0.0
channels-redis==4.0.0
Django==5.1.1
django-cors-headers==3.14.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
google-api-python-client==2.160.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pillow==11.1.0
psycopg2-binary==2.9.8
pyotp==2.9.0
redis==5.0.0
requests==2.32.3
